requirements.txt
kafka-python
pyflink
openai
chromadb
sentence-transformers
psycopg2-binary
